requirements.txt
torch
torchvision
gradio
Pillow
mlflow
scikit-learn
matplotlib